Moved proxies tool
