+++ Arma3ObjectBuilder/props/scene.py
@@ -11,17 +11,6 @@ class A3OB_PG_proxy_access(bpy.types.PropertyGroup):
     proxies_index: bpy.props.IntProperty(name="Selection Index")
 
 
-class A3OB_PG_lod_object(bpy.types.PropertyGroup):
-    name: bpy.props.StringProperty(name="Object Name")
-    lod: bpy.props.StringProperty(name="LOD type")
-    enabled: bpy.props.BoolProperty(name="Enabled")
-
-
-class A3OB_PG_proxies(bpy.types.PropertyGroup):
-    lod_objects: bpy.props.CollectionProperty(type=A3OB_PG_lod_object)
-    lod_objects_index: bpy.props.IntProperty(name="Selection Index")
-
-
 class A3OB_PG_common_data_item(bpy.types.PropertyGroup):
     name: bpy.props.StringProperty(name="Name", description="Descriptive name of the common item")
     value: bpy.props.StringProperty(name="Value", description="Value of the common item")
@@ -36,8 +25,6 @@ class A3OB_PG_common_data(bpy.types.PropertyGroup):
 classes = (
     A3OB_PG_proxy_access_item,
     A3OB_PG_proxy_access,
-    A3OB_PG_lod_object,
-    A3OB_PG_proxies,
     A3OB_PG_common_data_item,
     A3OB_PG_common_data
 )
@@ -47,7 +34,6 @@ def register():
     for cls in classes:
         bpy.utils.register_class(cls)
         
-    bpy.types.Scene.a3ob_proxies = bpy.props.PointerProperty(type=A3OB_PG_proxies)
     bpy.types.Scene.a3ob_commons = bpy.props.PointerProperty(type=A3OB_PG_common_data)
     bpy.types.Scene.a3ob_proxy_access = bpy.props.PointerProperty(type=A3OB_PG_proxy_access)
     
@@ -57,7 +43,6 @@ def register():
 def unregister():
     del bpy.types.Scene.a3ob_proxy_access
     del bpy.types.Scene.a3ob_commons
-    del bpy.types.Scene.a3ob_proxies
     
     for cls in reversed(classes):
         bpy.utils.unregister_class(cls)
