+++ Arma3ObjectBuilder/ui/__init__.py
@@ -8,6 +8,5 @@ from . import import_export_paa
 from . import props_action
 from . import props_material
 from . import props_object_mesh
-from . import tool_proxies
 from . import tool_scripts
 from . import tool_utilities
