+++ Arma3ObjectBuilder/utilities/__init__.py
@@ -1,3 +1,2 @@
 # In order of dependency
-from . import proxy
 from . import structure
